refactor(logic): split orders_application into parts for #1531
Extract work-phase application and completed-work handlers into part files
so each compilation unit stays under the 800-line guideline. Public API and
export path unchanged.

Refs #1531

Made-with: Cursor
