+++ packages/colonizethis_logic/lib/src/orders/orders_application_work_phase.dart
@@ -1,0 +1,390 @@
+part of 'orders_application.dart';
+
+void _runWorkPhase(
+  _BuildWorkState state,
+  void Function(_BuildWorkState, Unit, String) applyExploreCompletion,
+  void Function(
+    _BuildWorkState,
+    Unit,
+    CurrentWork,
+    List<Province> Function(),
+    void Function(List<Province>),
+  )
+  applyCompletedWorkTarget,
+) {
+  final workOrders = state.workOrders;
+  final tileState = state.tileState;
+  final oldUnitsById = state.oldUnitsById;
+  final newUnitsById = state.newUnitsById;
+  final purchasedTilesByTileKey = state.purchasedTilesByTileKey;
+
+  for (final player in state.game.players) {
+    var stockpile = player.stockpile;
+    var workers = player.workerPool;
+    var treasury = player.treasury;
+
+    Unit? lookupUnit(String unitId) =>
+        oldUnitsById[unitId] ?? newUnitsById[unitId];
+
+    void updateUnit(String unitId, Unit updated) {
+      if (oldUnitsById.containsKey(unitId)) {
+        oldUnitsById[unitId] = updated;
+      } else {
+        newUnitsById[unitId] = updated;
+      }
+    }
+
+    String regionForUnit(String unitId) =>
+        oldUnitsById.containsKey(unitId) ? kRegionOldWorld : kRegionNewWorld;
+
+    Province? provinceById(String id) =>
+        tryGetProvince(state.game.worldState, id);
+
+    bool canAffordMaterialCost(WorkOrderCost cost) {
+      for (final e in cost.entries) {
+        if (stockpile.quantityOf(e.key) < e.value) return false;
+      }
+      return true;
+    }
+
+    void deductMaterialCost(WorkOrderCost cost) {
+      for (final e in cost.entries) {
+        stockpile = stockpile.applyDelta(e.key, -e.value);
+      }
+    }
+
+    for (final order in workOrders[player.id] ?? const []) {
+      final u = lookupUnit(order.unitId);
+      if (u == null) continue;
+      final targetTileKey = order.targetTileKey;
+      final hasValidTarget = targetTileKey.isNotEmpty;
+
+      // Configuration for standard work order targets that use material cost.
+      // Reduces duplication in WorkOrder application by centralizing validation,
+      // cost computation, and unit update logic.
+      // Special targets (purchase_land, prospect, explore, steal_tech, counter_spy)
+      // are handled separately due to their unique logic.
+      ({
+        String target,
+        bool Function(String) allowedForUnitType,
+        WorkOrderCost? Function() costFn,
+        int Function() totalTurnsFn,
+      })
+      workTargetConfig(String target) {
+        switch (target) {
+          case 'build_improvement':
+            return (
+              target: target,
+              allowedForUnitType: (t) =>
+                  isWorkOrderTargetAllowedForUnitType(t, target),
+              costFn: () => workOrderMaterialCost(
+                target,
+                improvementLevel: tileState.improvementLevel(targetTileKey),
+              ),
+              totalTurnsFn: () => totalTurnsForWork(
+                target,
+                improvementLevel: tileState.improvementLevel(targetTileKey),
+              ),
+            );
+          case 'build_road':
+            return (
+              target: target,
+              allowedForUnitType: (t) =>
+                  isWorkOrderTargetAllowedForUnitType(t, target),
+              costFn: () => workOrderMaterialCost(target),
+              totalTurnsFn: () => totalTurnsForWork(target),
+            );
+          case 'build_port':
+            return (
+              target: target,
+              allowedForUnitType: (t) =>
+                  isWorkOrderTargetAllowedForUnitType(t, target),
+              costFn: () => workOrderMaterialCost(target),
+              totalTurnsFn: () => totalTurnsForWork(target),
+            );
+          case 'build_fort':
+            return (
+              target: target,
+              allowedForUnitType: (t) =>
+                  isWorkOrderTargetAllowedForUnitType(t, target),
+              costFn: () {
+                final prov = provinceById(u.locationProvinceId);
+                final fortLevel = prov?.fortLevel ?? 0;
+                return workOrderMaterialCost(target, fortLevel: fortLevel);
+              },
+              totalTurnsFn: () {
+                final prov = provinceById(u.locationProvinceId);
+                final fortLevel = prov?.fortLevel ?? 0;
+                return totalTurnsForWork(target, fortLevel: fortLevel);
+              },
+            );
+          case 'build_rail':
+            return (
+              target: target,
+              allowedForUnitType: (t) =>
+                  isWorkOrderTargetAllowedForUnitType(t, target),
+              costFn: () => workOrderMaterialCost(target),
+              totalTurnsFn: () => totalTurnsForWork(target),
+            );
+          case 'upgrade_town':
+            return (
+              target: target,
+              allowedForUnitType: (t) =>
+                  isWorkOrderTargetAllowedForUnitType(t, target),
+              costFn: () => workOrderMaterialCost(target),
+              totalTurnsFn: () => totalTurnsForWork(target),
+            );
+          default:
+            // Fall through to individual handling for non-standard targets
+            return (
+              target: target,
+              allowedForUnitType: (_) => false,
+              costFn: () => null,
+              totalTurnsFn: () => 1,
+            );
+        }
+      }
+
+      // Applies a standard work order using the config dispatch.
+      // Returns true if the order was applied, false otherwise.
+      bool applyStandardWorkOrder(String orderTarget) {
+        if (u.currentWork != null) return false;
+        if (!hasValidTarget) return false;
+
+        final config = workTargetConfig(orderTarget);
+        if (!config.allowedForUnitType(u.type)) return false;
+
+        final cost = config.costFn();
+        if (cost == null) return false;
+        if (!canAffordMaterialCost(cost)) return false;
+
+        deductMaterialCost(cost);
+        final totalTurns = config.totalTurnsFn();
+
+        _log.d(
+          'work order accepted and assigned unit=${order.unitId} target=$orderTarget targetTileKey=$targetTileKey totalTurns=$totalTurns',
+        );
+        updateUnit(
+          order.unitId,
+          u.copyWith(
+            status: UnitStatus.working,
+            tileKey: targetTileKey,
+            currentWork: CurrentWork(
+              workTarget: orderTarget,
+              tileKey: targetTileKey,
+              totalTurns: totalTurns,
+              remainingTurns: totalTurns,
+            ),
+          ),
+        );
+        return true;
+      }
+
+      if (order.target == kWorkTargetPurchaseLand &&
+          isWorkOrderTargetAllowedForUnitType(
+            u.type,
+            kWorkTargetPurchaseLand,
+          ) &&
+          hasValidTarget) {
+        // SPEC/game/diplomacy.md (GP–Minor/Tribe Rules): purchase_land requires an Embassy
+        // with the Minor/Tribe and the buyer must not be at war with that faction.
+        final resourceId =
+            state.game.worldState.resourceByTileKey[targetTileKey];
+        if (resourceId != null) {
+          final provinceId =
+              Unit.provinceIdFromTileKey(targetTileKey) ?? u.locationProvinceId;
+          final province = provinceById(provinceId);
+          final ownerId = province?.ownerId;
+          if (ownerId == null) {
+            continue;
+          }
+
+          final hasEmbassy = state.game.overtureStates.any(
+            (o) => o.gpId == player.id && o.targetId == ownerId && o.hasEmbassy,
+          );
+          if (!hasEmbassy) {
+            continue;
+          }
+
+          final atWar = state.game.diplomacyRelations.any((rel) {
+            final ids = {rel.factionId1, rel.factionId2};
+            return ids.contains(player.id) &&
+                ids.contains(ownerId) &&
+                rel.atWar;
+          });
+          if (atWar) {
+            continue;
+          }
+
+          final cost = purchaseLandCost(resourceId);
+          if (treasury >= cost) {
+            // First purchaser wins; tile can only be owned by one GP. SPEC/civilian-units.md.
+            if (!purchasedTilesByTileKey.containsKey(targetTileKey)) {
+              treasury -= cost;
+              purchasedTilesByTileKey[targetTileKey] = player.id;
+            }
+          }
+        }
+        continue;
+      }
+
+      if (order.target == kWorkTargetStealTech &&
+          isWorkOrderTargetAllowedForUnitType(u.type, kWorkTargetStealTech) &&
+          u.currentWork == null &&
+          hasValidTarget) {
+        final totalTurns = totalTurnsForWork(kWorkTargetStealTech);
+        _log.d(
+          'work order accepted and assigned unit=${order.unitId} target=steal_tech targetTileKey=$targetTileKey totalTurns=$totalTurns',
+        );
+        updateUnit(
+          order.unitId,
+          u.copyWith(
+            status: UnitStatus.working,
+            tileKey: targetTileKey,
+            currentWork: CurrentWork(
+              workTarget: kWorkTargetStealTech,
+              tileKey: targetTileKey,
+              totalTurns: totalTurns,
+              remainingTurns: totalTurns,
+            ),
+          ),
+        );
+        continue;
+      }
+
+      if (order.target == kWorkTargetCounterSpy &&
+          isWorkOrderTargetAllowedForUnitType(u.type, kWorkTargetCounterSpy) &&
+          u.currentWork == null &&
+          hasValidTarget) {
+        final totalTurns = totalTurnsForWork(kWorkTargetCounterSpy);
+        _log.d(
+          'work order accepted and assigned unit=${order.unitId} target=counter_spy targetTileKey=$targetTileKey totalTurns=$totalTurns',
+        );
+        updateUnit(
+          order.unitId,
+          u.copyWith(
+            status: UnitStatus.working,
+            tileKey: targetTileKey,
+            currentWork: CurrentWork(
+              workTarget: kWorkTargetCounterSpy,
+              tileKey: targetTileKey,
+              totalTurns: totalTurns,
+              remainingTurns: 1,
+            ),
+          ),
+        );
+        continue;
+      }
+
+      if (order.target == kWorkTargetProspect &&
+          hasValidTarget &&
+          isExplorerUnit(u.type) &&
+          isMineralEligibleTile(
+            state.game,
+            state.tileMapByRegion,
+            targetTileKey,
+          )) {
+        final existing =
+            state.game.worldState.playerProspectedTiles[player.id] ?? const {};
+        final newProspected = Set<String>.from(existing)..add(targetTileKey);
+        state.game = state.game.copyWith(
+          worldState: state.game.worldState.copyWith(
+            playerProspectedTiles: {
+              ...state.game.worldState.playerProspectedTiles,
+              player.id: newProspected,
+            },
+          ),
+        );
+      }
+      if (order.target == 'build_improvement') {
+        if (applyStandardWorkOrder('build_improvement')) continue;
+      }
+      if (order.target == kWorkTargetExplore &&
+          isExplorerUnit(u.type) &&
+          u.currentWork == null &&
+          hasValidTarget) {
+        final regionId = regionForUnit(order.unitId);
+        final provinceId =
+            Unit.provinceIdFromTileKey(targetTileKey) ?? u.locationProvinceId;
+        final byProvince =
+            state.game.worldState.tileKeysByRegionAndProvince[regionId];
+        final tilesInP = byProvince?[provinceId]?.length ?? 0;
+        if (tilesInP > 0 && byProvince != null && byProvince.isNotEmpty) {
+          var maxTiles = 0;
+          for (final list in byProvince.values) {
+            if (list.length > maxTiles) maxTiles = list.length;
+          }
+          if (maxTiles < 1) maxTiles = 1;
+          final totalTurns = (3 * tilesInP / maxTiles).ceil().clamp(1, 999);
+          _log.d(
+            'work order accepted and assigned unit=${order.unitId} target=explore targetTileKey=$targetTileKey totalTurns=$totalTurns',
+          );
+          updateUnit(
+            order.unitId,
+            u.copyWith(
+              status: UnitStatus.working,
+              tileKey: targetTileKey,
+              currentWork: CurrentWork(
+                workTarget: kWorkTargetExplore,
+                tileKey: targetTileKey,
+                totalTurns: totalTurns,
+                remainingTurns: totalTurns,
+              ),
+            ),
+          );
+          continue;
+        }
+      }
+      final workTarget = order.target;
+      if (workTarget == 'build_road') {
+        if (applyStandardWorkOrder('build_road')) continue;
+      }
+      if (workTarget == 'build_port') {
+        if (applyStandardWorkOrder('build_port')) continue;
+      }
+      if (workTarget == 'build_fort') {
+        final prov = provinceById(u.locationProvinceId);
+        final fortLevel = prov?.fortLevel ?? 0;
+        if (fortLevel == 1 &&
+            player.techUnlocked?['mine_engineering'] != true) {
+          _log.d(
+            'build_fort skipped - Mine Engineering required for fort level 2',
+          );
+          continue;
+        }
+        if (fortLevel == 2 && player.techUnlocked?['modern_forts'] != true) {
+          _log.d('build_fort skipped - Modern Forts required for fort level 3');
+          continue;
+        }
+        if (applyStandardWorkOrder('build_fort')) continue;
+      }
+      if (workTarget == 'build_rail') {
+        final terrain = terrainTypeForTileKey(
+          state.tileMapByRegion,
+          targetTileKey,
+        );
+        final railReason = rejectionReasonForBuildRailOrder(
+          techUnlocked: player.techUnlocked,
+          roadLevel: tileState.roadLevel(targetTileKey),
+          terrain: terrain,
+        );
+        if (railReason != null) {
+          _log.d('build_rail skipped - $railReason');
+          continue;
+        }
+        if (applyStandardWorkOrder('build_rail')) continue;
+      }
+      if (workTarget == 'upgrade_town') {
+        if (applyStandardWorkOrder('upgrade_town')) continue;
+      }
+    }
+
+    state.updatedPlayers.add(
+      player.copyWith(
+        stockpile: stockpile,
+        workerPool: workers,
+        treasury: treasury,
+      ),
+    );
+  }
+}
